Make login time configurable
