+++ asreview/webapp/api/auth.py
@@ -13,21 +13,15 @@
 # limitations under the License.
 
 
-import datetime as dt
-from pathlib import Path
-
 from flask import Blueprint
 from flask import current_app
 from flask import jsonify
-from flask import render_template_string
 from flask import request
 from flask import redirect
 from flask_login import current_user
 from flask_login import login_required
-from flask_login import login_user
 from flask_login import logout_user
-from flask_mail import Mail
-from flask_mail import Message
+
 from sqlalchemy import and_
 from sqlalchemy import or_
 from sqlalchemy.exc import IntegrityError
@@ -38,73 +32,13 @@ from asreview.webapp import DB
 from asreview.webapp.authentication.models import User
 from asreview.webapp.authentication.oauth_handler import OAuthHandler
 
+from asreview.webapp.authentication.utils import has_email_configuration
+from asreview.webapp.authentication.utils import perform_login_user
+from asreview.webapp.authentication.utils import send_forgot_password_email
+from asreview.webapp.authentication.utils import send_confirm_account_email
+
+
 bp = Blueprint("auth", __name__, url_prefix="/auth")
-
-
-def _has_email_configuration(app):
-    return all(
-        [
-            app.config.get("MAIL_SERVER", False),
-            app.config.get("MAIL_USERNAME", False),
-            app.config.get("MAIL_PASSWORD", False),
-        ]
-    )
-
-
-def perform_login_user(user):
-    """Helper function to login a user"""
-    return login_user(user, remember=True, duration=dt.timedelta(days=31))
-
-
-# TODO: not sure if this file is the right place for this function
-def send_forgot_password_email(user, request, cur_app):
-    # do not send email in test environment
-    if not cur_app.testing:
-        # get necessary information out of user object
-        name = user.name or "ASReview user"
-        # get url of front-end
-        root_url = request.headers.get("Origin")
-        # create url that will be used in the email
-        url = f"{root_url}/reset_password?user_id={user.id}&token={user.token}"
-        # create a mailer
-        mailer = Mail(cur_app)
-        # open templates as string and render
-        root_path = Path(cur_app.root_path)
-        with open(root_path / "templates" / "emails" / "forgot_password.html") as f:
-            html_text = render_template_string(f.read(), name=name, url=url)
-        with open(root_path / "templates" / "emails" / "forgot_password.txt") as f:
-            txt_text = render_template_string(f.read(), name=name, url=url)
-        # create message
-        msg = Message("ASReview: forgot password", recipients=[user.email])
-        msg.body = txt_text
-        msg.html = html_text
-        return mailer.send(msg)
-
-
-# TODO: not sure if this file is the right place for this function
-def send_confirm_account_email(user, request, cur_app):
-    # do not send email in test environment
-    if not cur_app.testing:
-        # get necessary information out of user object
-        name = user.name or "ASReview user"
-        # get url of front-end
-        root_url = request.headers.get("Origin")
-        # create url that will be used in the email
-        url = f"{root_url}/confirm_account?user_id={user.id}&token={user.token}"
-        # create a mailer
-        mailer = Mail(cur_app)
-        # open templates as string and render
-        root_path = Path(cur_app.root_path)
-        with open(root_path / "templates" / "emails" / "confirm_account.html") as f:
-            html_text = render_template_string(f.read(), name=name, url=url)
-        with open(root_path / "templates" / "emails" / "confirm_account.txt") as f:
-            txt_text = render_template_string(f.read(), name=name, url=url)
-        # create message
-        msg = Message("ASReview: please confirm your account", recipients=[user.email])
-        msg.body = txt_text
-        msg.html = html_text
-        return mailer.send(msg)
-
 
 # ------------------
 #      ROUTES
@@ -130,7 +64,7 @@ def signin():
     else:
         # user exists and is confirmed: verify password
         if user.verify_password(password):
-            if perform_login_user(user):
+            if perform_login_user(user, current_app):
                 result = (
                     200,
                     {
@@ -301,7 +235,7 @@ def get_profile():
 
 @bp.route("/forgot_password", methods=["POST"])
 def forgot_password():
-    if _has_email_configuration(current_app):
+    if has_email_configuration(current_app):
         # get email address from request
         email_address = request.form.get("email", "").strip()
 
@@ -342,7 +276,7 @@ def forgot_password():
 @bp.route("/reset_password", methods=["POST"])
 def reset_password():
     """Resests password of user"""
-    if _has_email_configuration(current_app):
+    if has_email_configuration(current_app):
         new_password = request.form.get("password", "").strip()
         token = request.form.get("token", "").strip()
         user_id = request.form.get("user_id", "0").strip()
@@ -490,7 +424,7 @@ def oauth_callback():
                 return jsonify({"message": message}), 409
 
         # log in the existing/created user immediately
-        logged_in = perform_login_user(user)
+        logged_in = perform_login_user(user, current_app)
         result = (
             200,
             {
